# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "ApprovalRequest": {
        "description": "Hash-bound approval request.",
        "properties": {
          "budget": {
            "additionalProperties": true,
            "title": "Budget",
            "type": "object"
          },
          "expected_content_hash": {
            "maxLength": 64,
            "minLength": 64,
            "title": "Expected Content Hash",
            "type": "string"
          }
        },
        "required": [
          "expected_content_hash"
        ],
        "title": "ApprovalRequest",
        "type": "object"
      },
      "ApprovalResponse": {
        "description": "Durable identifiers created by approval.",
        "properties": {
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "task_id": {
            "format": "uuid",
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "task_id",
          "job_id"
        ],
        "title": "ApprovalResponse",
        "type": "object"
      },
      "CheckpointRequest": {
        "properties": {
          "checkpoint": {
            "additionalProperties": true,
            "title": "Checkpoint",
            "type": "object"
          },
          "generation": {
            "minimum": 1.0,
            "title": "Generation",
            "type": "integer"
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "worker_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Worker Id",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "worker_id",
          "generation"
        ],
        "title": "CheckpointRequest",
        "type": "object"
      },
      "CompleteRequest": {
        "properties": {
          "generation": {
            "minimum": 1.0,
            "title": "Generation",
            "type": "integer"
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "result_refs": {
            "additionalProperties": true,
            "title": "Result Refs",
            "type": "object"
          },
          "worker_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Worker Id",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "worker_id",
          "generation"
        ],
        "title": "CompleteRequest",
        "type": "object"
      },
      "DependencyStatus": {
        "description": "Sanitized status for one external dependency.",
        "properties": {
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "DependencyStatus",
        "type": "object"
      },
      "FailRequest": {
        "properties": {
          "error_class": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Error Class",
            "type": "string"
          },
          "generation": {
            "minimum": 1.0,
            "title": "Generation",
            "type": "integer"
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "retry_after_seconds": {
            "anyOf": [
              {
                "maximum": 86400.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry After Seconds"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          },
          "worker_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Worker Id",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "worker_id",
          "generation",
          "error_class"
        ],
        "title": "FailRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HealthResponse": {
        "description": "Stable process-health response.",
        "properties": {
          "service": {
            "title": "Service",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "service",
          "status",
          "version"
        ],
        "title": "HealthResponse",
        "type": "object"
      },
      "HeartbeatRequest": {
        "properties": {
          "generation": {
            "minimum": 1.0,
            "title": "Generation",
            "type": "integer"
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "lease_seconds": {
            "default": 60,
            "maximum": 3600.0,
            "minimum": 1.0,
            "title": "Lease Seconds",
            "type": "integer"
          },
          "worker_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Worker Id",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "worker_id",
          "generation"
        ],
        "title": "HeartbeatRequest",
        "type": "object"
      },
      "ReadinessResponse": {
        "description": "Stable readiness response used by supervisors and dashboards.",
        "properties": {
          "dependencies": {
            "additionalProperties": {
              "$ref": "#/components/schemas/DependencyStatus"
            },
            "title": "Dependencies",
            "type": "object"
          },
          "service": {
            "title": "Service",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "service",
          "status",
          "dependencies"
        ],
        "title": "ReadinessResponse",
        "type": "object"
      },
      "RunCancelRequest": {
        "properties": {
          "reason": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "RunCancelRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WorkerClaimRequest": {
        "properties": {
          "lease_seconds": {
            "default": 60,
            "maximum": 3600.0,
            "minimum": -3600.0,
            "title": "Lease Seconds",
            "type": "integer"
          },
          "worker_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Worker Id",
            "type": "string"
          }
        },
        "required": [
          "worker_id"
        ],
        "title": "WorkerClaimRequest",
        "type": "object"
      },
      "WorkerLeaseResponse": {
        "properties": {
          "attempt_id": {
            "format": "uuid",
            "title": "Attempt Id",
            "type": "string"
          },
          "cancellation_epoch": {
            "title": "Cancellation Epoch",
            "type": "integer"
          },
          "generation": {
            "title": "Generation",
            "type": "integer"
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "lease_until": {
            "format": "date-time",
            "title": "Lease Until",
            "type": "string"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "task_id": {
            "format": "uuid",
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "task_id",
          "run_id",
          "attempt_id",
          "generation",
          "cancellation_epoch",
          "lease_until"
        ],
        "title": "WorkerLeaseResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Ebook Factory API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/health": {
      "get": {
        "description": "Return process health without requiring external dependencies.",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health",
        "tags": [
          "operations"
        ]
      }
    },
    "/private/runs/{run_id}/cancel": {
      "post": {
        "operationId": "worker_cancel_private_runs__run_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Ebook-Worker-Token",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Ebook-Worker-Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCancelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "title": "Response Worker Cancel Private Runs  Run Id  Cancel Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Worker Cancel",
        "tags": [
          "private-worker"
        ]
      }
    },
    "/private/worker/checkpoint": {
      "post": {
        "operationId": "worker_checkpoint_private_worker_checkpoint_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Ebook-Worker-Token",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Ebook-Worker-Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckpointRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Worker Checkpoint",
        "tags": [
          "private-worker"
        ]
      }
    },
    "/private/worker/claim": {
      "post": {
        "operationId": "worker_claim_private_worker_claim_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Ebook-Worker-Token",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Ebook-Worker-Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkerClaimRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/WorkerLeaseResponse"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Worker Claim Private Worker Claim Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Worker Claim",
        "tags": [
          "private-worker"
        ]
      }
    },
    "/private/worker/complete": {
      "post": {
        "operationId": "worker_complete_private_worker_complete_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Ebook-Worker-Token",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Ebook-Worker-Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Worker Complete",
        "tags": [
          "private-worker"
        ]
      }
    },
    "/private/worker/fail": {
      "post": {
        "operationId": "worker_fail_private_worker_fail_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Ebook-Worker-Token",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Ebook-Worker-Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Worker Fail",
        "tags": [
          "private-worker"
        ]
      }
    },
    "/private/worker/heartbeat": {
      "post": {
        "operationId": "worker_heartbeat_private_worker_heartbeat_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Ebook-Worker-Token",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Ebook-Worker-Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "format": "date-time",
                    "type": "string"
                  },
                  "title": "Response Worker Heartbeat Private Worker Heartbeat Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Worker Heartbeat",
        "tags": [
          "private-worker"
        ]
      }
    },
    "/projects/{project_id}/briefs/{brief_id}/approve": {
      "post": {
        "operationId": "approve_projects__project_id__briefs__brief_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "brief_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Brief Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve"
      }
    },
    "/projects/{project_id}/events": {
      "get": {
        "operationId": "events_projects__project_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "default": 0,
              "title": "After",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Events Projects  Project Id  Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Events"
      }
    },
    "/ready": {
      "get": {
        "description": "Return dependency readiness with sanitized database status.",
        "operationId": "ready_ready_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadinessResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Ready",
        "tags": [
          "operations"
        ]
      }
    }
  }
}